fix: apply service account to spec
The current structure applies the `serviceAccountName` to the container definition, but the reference documentation for the container definition doesn't seem to support a `serviceAccountName` field. Instead, it looks like this is supposed to be defined on the `spec` itself. This looks like it's causing my attempts to use this chart with a specific service account (rather than the `default` SA) to fail.