Set GOMAXPROCS based on CPU request
This prevents Go from detecting the total number of CPUs on the host and scheduling that many threads. If Go tries to run more threads than cores it has available, it thrashes and the performance tanks. So if we only have 2 guaranteed CPU cores allocated to us out of the 8 cores on the host, we need to set GOMAXPROCS=2.